Fixed the data type

+++ backend/api/models/overall_score.go
@@ -1,6 +1,6 @@
 package models
 
 type OverallScore struct {
-	Score          float64 `json:"score"`
-	Interpretation string  `json:"interpretation"`
+	Score          int32  `json:"score"`
+	Interpretation string `json:"interpretation"`
 }
